+++ dtool/src/parser-inc/winsock2.h
@@ -2,7 +2,9 @@
 #define _WINSOCK2API_
 #define _WINSOCKAPI_   /* Prevent inclusion of winsock.h in windows.h */
 
-typedef int SOCKET;
+#include <wtypes.h>
+
+typedef unsigned long SOCKET;
 
 struct sockaddr_in;
